refactor: move lifecycle helper storage contracts (#3595)
